pyproject.toml
[tool.poetry]
name = "pollloop"
version = "0.1.0"
description = ""
authors = ["Hyunminmax <hyunmin.shin81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2024.12.1"
django-seed = "^0.3.1"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
